backend/requirements.txt
earthengine-api
fastapi
uvicorn
pydantic
numpy
scikit-learn
Pillow
rasterio
